refactor: move all transformer benchmarks to dedicated *_bench_test.go files
