+++ transformers/geoip_test.go
@@ -188,47 +188,3 @@ func TestGeoIP_LookupCoordinate(t *testing.T) {
 		t.Errorf("longitude invalid want: 0 got: %f", geoInfo.Longitude)
 	}
 }
-
-func BenchmarkGeoIP_Lookup(b *testing.B) {
-	config := pkgconfig.GetFakeConfigTransformers()
-	config.GeoIP.Enable = true
-	config.GeoIP.DBCountryFile = "../tests/testsdata/GeoLite2-Country.mmdb"
-	config.GeoIP.DBASNFile = "../tests/testsdata/GeoLite2-ASN.mmdb"
-
-	outChans := []chan *dnsutils.DNSMessage{}
-	geoip := NewDNSGeoIPTransform(config, logger.New(false), "test", 0, outChans)
-	if err := geoip.Open(); err != nil {
-		b.Fatalf("geoip init failed: %v", err)
-	}
-	defer geoip.Close()
-
-	ip := "83.112.146.176"
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		_, _ = geoip.Lookup(ip)
-	}
-}
-
-func BenchmarkGeoIP_Lookup_ECS(b *testing.B) {
-	config := pkgconfig.GetFakeConfigTransformers()
-	config.GeoIP.Enable = true
-	config.GeoIP.DBCountryFile = "../tests/testsdata/GeoLite2-Country.mmdb"
-	config.GeoIP.LookupECS = true
-
-	outChans := []chan *dnsutils.DNSMessage{}
-	geoip := NewDNSGeoIPTransform(config, logger.New(false), "test", 0, outChans)
-	if err := geoip.Open(); err != nil {
-		b.Fatalf("geoip init failed: %v", err)
-	}
-	defer geoip.Close()
-
-	dm := dnsutils.GetFakeDNSMessage()
-	dm.NetworkInfo.QueryIP = "1.1.1.1"
-	dm.EDNS.Options = append(dm.EDNS.Options, dnsutils.DNSOption{Code: 8, Name: "CSUBNET", Data: "83.112.146.176/32"})
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		_, _ = geoip.geoipTransform(&dm)
-	}
-}
