+++ transformers/latency_test.go
@@ -1,8 +1,6 @@
 package transformers
 
 import (
-	"runtime"
-	"sync"
 	"testing"
 	"time"
 
@@ -146,71 +144,4 @@ func Test_HashQueries_Expire(t *testing.T) {
 	if ok {
 		t.Errorf("key/value always in map!")
 	}
-}
-
-// Bench
-func Benchmark_HashQueries_Set(b *testing.B) {
-	mapexpire := NewHashQueries(10 * time.Second)
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		mapexpire.Set(uint64(i), int64(i))
-	}
-}
-
-func Benchmark_HashQueries_Delete(b *testing.B) {
-	mapexpire := NewHashQueries(60 * time.Second)
-
-	for i := 0; i < b.N; i++ {
-		mapexpire.Set(uint64(i), int64(i))
-	}
-
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		mapexpire.Delete(uint64(i))
-	}
-}
-
-func Benchmark_HashQueries_Get(b *testing.B) {
-	mapexpire := NewHashQueries(60 * time.Second)
-
-	for i := 0; i < b.N; i++ {
-		mapexpire.Set(uint64(i), int64(i))
-	}
-
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		_, ok := mapexpire.Get(uint64(i))
-		if !ok {
-			break
-		}
-
-	}
-}
-
-func Benchmark_HashQueries_ConcurrentGet(b *testing.B) {
-	mapexpire := NewHashQueries(60 * time.Second)
-	for i := 0; i < b.N; i++ {
-		mapexpire.Set(uint64(i), int64(i))
-	}
-
-	var wg sync.WaitGroup
-	b.ResetTimer()
-
-	for wc := 0; wc < runtime.NumCPU(); wc++ {
-		wg.Add(1)
-		go func(n int) {
-			defer wg.Done()
-			for i := 0; i < n; i++ {
-				_, ok := mapexpire.Get(uint64(i))
-				if !ok {
-					break
-				}
-			}
-		}(b.N)
-	}
-
-	wg.Wait()
 }
